# Jpeg decoder mcu size fix