### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
This PR fixes a bug described in https://github.com/SixLabors/ImageSharp/pull/1816#issuecomment-965916098.

Basically, it removes dependency on the 'first component' during decoding because jpegs can have lower luma sampling than chroma. This bug can be spotted with image from current test image base - [jpeg422.jpg](https://github.com/SixLabors/ImageSharp/blob/master/tests/Images/Input/Jpg/baseline/jpeg422.jpg). This is a decoded image from current master code:
![jpeg422_MASTER_BUG](https://user-images.githubusercontent.com/20967409/142748379-92a653ca-1410-425f-b27c-ae8fbc71d9f3.png)

Current PR produces almost ideal image - diff is 0.0013% compared to PhotoShop png reference:

```
MASTER
Total difference: 19,7988%
[Δ(0,-20560,0,0) @ (0,0)];
[Δ(0,-20560,0,0) @ (1,0)];
[Δ(0,-20303,-257,0) @ (2,0)];
[Δ(0,-20046,0,0) @ (3,0)];
[Δ(0,-19789,-257,0) @ (4,0)]...
```
```
PR
Total difference: 0,0013%
[Δ(0,-257,257,0) @ (18,2)];
[Δ(0,-257,0,0) @ (19,3)];
[Δ(0,-257,0,0) @ (23,7)];
[Δ(0,257,0,0) @ (32,8)];
[Δ(0,0,-257,0) @ (56,8)]...
```

I've also added this image to the decoder test suit.