@br3aker 2021-11-21
Fresh benchmarks

Master
|                              Method |      Mean |     Error |    StdDev |
|------------------------------------ |----------:|----------:|----------:|
|        'Baseline 4:4:4 Interleaved' | 11.712 ms | 0.0454 ms | 0.0355 ms |
|        'Baseline 4:2:0 Interleaved' |  8.931 ms | 0.0513 ms | 0.0455 ms |
|        'Baseline 4:0:0 (grayscale)' |  1.639 ms | 0.0045 ms | 0.0037 ms |
| 'Progressive 4:2:0 Non-Interleaved' | 13.554 ms | 0.0711 ms | 0.0665 ms |


PR
|                              Method |      Mean |     Error |    StdDev |
|------------------------------------ |----------:|----------:|----------:|
|        'Baseline 4:4:4 Interleaved' | 11.781 ms | 0.0737 ms | 0.0654 ms |
|        'Baseline 4:2:0 Interleaved' |  8.688 ms | 0.0345 ms | 0.0306 ms |
|        'Baseline 4:0:0 (grayscale)' |  1.643 ms | 0.0092 ms | 0.0086 ms |
| 'Progressive 4:2:0 Non-Interleaved' | 13.770 ms | 0.0928 ms | 0.0823 ms |

While progressive 420 is indeed slower by almost 1.5% baseline 420 is faster by 2.7% without any particular reasons. I believe it's a noise, performance shouldn't be affected by LINQ calls, especially by 4 element array.